Fix TypeScript error and linting issues: extract SortableTab to separate file
